+++ client/src/components/app.tsx
@@ -4,7 +4,7 @@ import { connect } from "react-redux";
 import { ThemeProvider as EmotionThemeProvider } from "@emotion/react";
 import { StylesProvider, ThemeProvider } from "@material-ui/core/styles";
 import { theme } from "./theme";
-import { BottomBanner } from "./BottomBanner";
+import BottomBanner from "./BottomBanner";
 
 import Controls from "./controls";
 import DatasetSelector from "./datasetSelector/datasetSelector";
